moved custom arrays to main class
