+++ src/index.ts
@@ -1,6 +1,5 @@
 import GenerateArray from './GenerateArray.ts';
-import CustomArray from "./CustomArray.ts";
 import ArrayUtils from "./ArrayUtils.ts";
 import GenerateArrayError from "./GenerateArrayError.ts";
 
-export { GenerateArray, CustomArray, ArrayUtils, GenerateArrayError };
+export { GenerateArray, ArrayUtils, GenerateArrayError };
